# frontend/src/api/generated/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "AttemptResponse": {
        "properties": {
          "item": {
            "$ref": "Attempt"
          },
          "success": {
            "const": true,
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "CreateAttemptRequest": {
        "properties": {
          "note": {
            "type": "string"
          },
          "workflow_hash": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CreateJobRequest": {
        "properties": {
          "kind": {
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object"
          },
          "project": {
            "type": "string"
          }
        },
        "required": [
          "project",
          "kind"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "description": "稳定错误码，代码里以此分支，不要解析文案",
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "success": {
            "const": false,
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "error",
          "code"
        ],
        "type": "object"
      },
      "FinishAttemptRequest": {
        "properties": {
          "note": {
            "type": "string"
          },
          "result_ref": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "JobListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "Job"
            },
            "type": "array"
          },
          "success": {
            "const": true,
            "type": "boolean"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "success",
          "items"
        ],
        "type": "object"
      },
      "JobResponse": {
        "properties": {
          "item": {
            "$ref": "Job"
          },
          "success": {
            "const": true,
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "JobStatsResponse": {
        "properties": {
          "stats": {
            "$ref": "JobStats"
          },
          "success": {
            "const": true,
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "stats"
        ],
        "type": "object"
      },
...
      "PendingUnitListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "PendingUnit"
            },
            "type": "array"
          },
          "success": {
            "const": true,
            "type": "boolean"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "success",
          "items"
        ],
        "type": "object"
      },
...
      "ReuseHintResponse": {
        "properties": {
          "reason": {
            "description": "为何不可复用；workflow_hash 不一致 / 产物缺失",
            "type": "string"
          },
          "reusable": {
            "type": "boolean"
          },
          "success": {
            "const": true,
            "type": "boolean"
          },
          "workflow_hash": {
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "UpdateAttemptRequest": {
        "properties": {
          "note": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "UpdateJobRequest": {
        "properties": {
          "note": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "本规格是**前端与后端的唯一契约**。前端客户端由 `scripts/generate_client.py` 从本文件生成，`scripts/check_generated_client.py` 在 CI 校验生成物是否过期。\n\n**手工改这个文件没有意义**：它是生成物，改了会被下一次生成覆盖。",
    "title": "漫剧锻造台 API",
    "version": "2026-10-07.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/jobs": {
      "get": {
        "operationId": "listJobs",
        "parameters": [
          {
            "description": "按状态过滤",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "作业列表（Job = 一次用户意图）",
        "tags": [
          "jobs"
        ],
        "x-contract-source": "manual"
      },
      "post": {
        "operationId": "createJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "参数非法"
          }
        },
        "summary": "创建作业",
        "tags": [
          "jobs"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/jobs/attempts/{attempt_id}/finish": {
      "post": {
        "operationId": "finishJobAttempt",
        "parameters": [
          {
            "description": "尝试 id",
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinishAttemptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "尝试不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "终态不可再变更"
          }
        },
        "summary": "结束尝试并结算",
        "tags": [
          "jobs"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/jobs/attempts/{attempt_id}/status": {
      "post": {
        "operationId": "updateAttemptStatus",
        "parameters": [
          {
            "description": "尝试 id",
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAttemptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "尝试不存在"
          }
        },
        "summary": "更新尝试状态",
        "tags": [
          "jobs"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/jobs/stats": {
      "get": {
        "operationId": "getJobStats",
        "parameters": [
          {
            "description": "按项目过滤",
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatsResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "作业统计（按状态聚合）",
        "tags": [
          "jobs"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "operationId": "getJob",
        "parameters": [
          {
            "description": "作业 id",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "作业不存在"
          }
        },
        "summary": "作业详情（含各次尝试）",
        "tags": [
          "jobs"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/jobs/{job_id}/attempts": {
      "post": {
        "operationId": "createJobAttempt",
        "parameters": [
          {
            "description": "作业 id",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttemptRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "作业不存在"
          }
        },
        "summary": "为作业开启一次尝试（Attempt = 一次执行）",
        "tags": [
          "jobs"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/jobs/{job_id}/pending-units": {
      "get": {
        "operationId": "listPendingUnits",
        "parameters": [
          {
            "description": "作业 id",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingUnitListResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "作业不存在"
          }
        },
        "summary": "待处理单元（断点续跑判据之一）",
        "tags": [
          "jobs"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/jobs/{job_id}/reuse-hint": {
      "get": {
        "operationId": "getJobReuseHint",
        "parameters": [
          {
            "description": "作业 id",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReuseHintResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "作业不存在"
          }
        },
        "summary": "免重渲提示（workflow_hash 命中可复用）",
        "tags": [
          "jobs"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/jobs/{job_id}/status": {
      "post": {
        "operationId": "updateJobStatus",
        "parameters": [
          {
            "description": "作业 id",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "作业不存在"
          }
        },
        "summary": "更新作业状态",
        "tags": [
          "jobs"
        ],
        "x-contract-source": "manual"
      }
    },
...
  },
  "servers": [
    {
      "description": "同源部署",
      "url": "/"
    }
  ],
...
}